增加category root category & parent category id

+++ src/main/java/myblog/resource/CategoryResource.java
@@ -18,14 +18,15 @@ public class CategoryResource {
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     public Response createCategory(Category category) {
-        if (!Category.isValidCategoryName(category.getCategory_name_cn())) {
-            throw new BadRequestException();
+        if (category.getCategory_name_cn() != null) {
+            if (!Category.isValidCategoryName(category.getCategory_name_cn())) {
+                throw new BadRequestException("Unexpected category name cn");
+            }
         }
-        if (!Category.isValidCategoryName(category.getCategory_name_en())) {
-            throw new BadRequestException();
-        }
-        if (!Category.isValidCategoryLevel(category.getCategory_level())) {
-            throw new BadRequestException();
+        if (category.getCategory_name_en() != null) {
+            if (!Category.isValidCategoryName(category.getCategory_name_en())) {
+                throw new BadRequestException("Unexpected category name en");
+            }
         }
         category.setCategory_enabled(true);
         category.setCategory_created_at(null);
@@ -35,7 +36,7 @@ public class CategoryResource {
         if (Category.isValidCategoryId(categoryId)) {
             return Response.created(URI.create("/categories/" + categoryId)).build();
         } else {
-            throw new InternalServerErrorException();
+            throw new InternalServerErrorException("Unexpected error");
         }
     }
 
@@ -43,7 +44,7 @@ public class CategoryResource {
     @Path("/{categoryId}")
     @Produces(MediaType.APPLICATION_JSON)
     public Response deleteCategory(@PathParam("categoryId") Integer categoryId) {
-        if (Category.isValidCategoryId(categoryId)) {
+        if (categoryId != null && Category.isValidCategoryId(categoryId)) {
             if (CategoryService.getCategoryById(categoryId) != null) {
                 if (CategoryService.deleteCategory(categoryId)) {
                     return Response.noContent().build();
@@ -51,10 +52,10 @@ public class CategoryResource {
                     throw new InternalServerErrorException();
                 }
             } else {
-                throw new NotFoundException();
+                throw new NotFoundException("Not found category");
             }
         } else {
-            throw new BadRequestException();
+            throw new BadRequestException("Unexpected category id");
         }
     }
 
@@ -63,22 +64,17 @@ public class CategoryResource {
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     public Response updateCategory(@PathParam("categoryId") Integer categoryId, Category category) {
-        if (!Category.isValidCategoryId(categoryId)) {
-            throw new BadRequestException();
+        if (categoryId == null || !Category.isValidCategoryId(categoryId)) {
+            throw new BadRequestException("Unexpected category id");
         }
         if (category.getCategory_name_cn() != null) {
             if (!Category.isValidCategoryName(category.getCategory_name_cn())) {
-                throw new BadRequestException();
+                throw new BadRequestException("Unexpected category name cn");
             }
         }
         if (category.getCategory_name_en() != null) {
             if (!Category.isValidCategoryName(category.getCategory_name_en())) {
-                throw new BadRequestException();
-            }
-        }
-        if (category.getCategory_level() != null) {
-            if (!Category.isValidCategoryLevel(category.getCategory_level())) {
-                throw new BadRequestException();
+                throw new BadRequestException("Unexpected category name en");
             }
         }
         category.setCategory_updated_at(null);
@@ -91,7 +87,7 @@ public class CategoryResource {
                 throw new InternalServerErrorException();
             }
         } else {
-            throw new NotFoundException();
+            throw new NotFoundException("Not found category");
         }
     }
 
@@ -99,40 +95,44 @@ public class CategoryResource {
     @Path("/{categoryId}")
     @Produces(MediaType.APPLICATION_JSON)
     public Category getCategoryById(@PathParam("categoryId") Integer categoryId) {
-        if (Category.isValidCategoryId(categoryId)) {
+        if (categoryId != null && Category.isValidCategoryId(categoryId)) {
             Category category = CategoryService.getCategoryById(categoryId);
 
             if (category != null) {
                 return category;
             } else {
-                throw new NotFoundException();
+                throw new NotFoundException("Not found category");
             }
         } else {
-            throw new BadRequestException();
+            throw new BadRequestException("Unexpected category id");
         }
     }
 
     @GET
     @Produces(MediaType.APPLICATION_JSON)
-    public List<Category> getCategoryList(@QueryParam("category_name") String categoryName,
-                                          @QueryParam("category_level") Integer categoryLevel,
-                                          @QueryParam("duration_start") String durationStart,
-                                          @QueryParam("duration_end") String durationEnd,
-                                          @QueryParam("category_enabled") Boolean categoryEnabled) {
+    public List<Category> getCategories(@QueryParam("category_parent_id") Integer categoryParentId,
+                                        @QueryParam("category_name") String categoryName,
+                                        @QueryParam("category_level") Integer categoryLevel,
+                                        @QueryParam("duration_begin") String durationBegin,
+                                        @QueryParam("duration_end") String durationEnd,
+                                        @QueryParam("category_enabled") Boolean categoryEnabled) {
         Category category = new Category();
-        if (Category.isValidCategoryName(categoryName)) {
+        if (categoryParentId != null) {
+            category.setCategory_parent_id(categoryParentId);
+        }
+        if (categoryName != null) {
             category.setCategory_name_en(categoryName);
         }
-        if (Category.isValidCategoryLevel(categoryLevel)) {
+        if (categoryLevel != null) {
             category.setCategory_level(categoryLevel);
         }
-        if (Category.isValidDurationBegin(durationStart)) {
-            category.setDuration_start(durationStart);
+        if (durationBegin != null) {
+            category.setDuration_begin(durationBegin);
         }
-        if (Category.isValidDurationEnd(durationEnd)) {
+        if (durationEnd != null) {
             category.setDuration_end(durationEnd);
         }
-        if (Category.isValidCategoryEnabled(categoryEnabled)) {
+        if (categoryEnabled != null) {
             category.setCategory_enabled(categoryEnabled);
         }
 
@@ -142,22 +142,23 @@ public class CategoryResource {
                 if (categories.size() > 0) {
                     return categories;
                 } else {
-                    throw new NotFoundException();
+                    throw new NotFoundException("Not found categories");
                 }
             } else {
                 throw new InternalServerErrorException();
             }
         } else {
-            throw new BadRequestException();
+            throw new BadRequestException("Unexpected category");
         }
     }
 
     @GET
     @Path("/list")
     @Produces(MediaType.APPLICATION_JSON)
-    public Pagination<Category> getCategoryList(@QueryParam("category_name") String categoryName,
+    public Pagination<Category> getCategoryList(@QueryParam("category_parent_id") Integer categoryParentId,
+                                                @QueryParam("category_name") String categoryName,
                                                 @QueryParam("category_level") Integer categoryLevel,
-                                                @QueryParam("duration_start") String durationStart,
+                                                @QueryParam("duration_begin") String durationBegin,
                                                 @QueryParam("duration_end") String durationEnd,
                                                 @QueryParam("category_enabled") Boolean categoryEnabled,
                                                 @QueryParam("limit") Integer limit,
@@ -165,21 +166,25 @@ public class CategoryResource {
                                                 @QueryParam("order_by") String orderBy,
                                                 @QueryParam("order_type") String orderType) {
         Category category = new Category();
-        if (Category.isValidCategoryName(categoryName)) {
+        if (categoryParentId != null) {
+            category.setCategory_parent_id(categoryParentId);
+        }
+        if (categoryName != null) {
             category.setCategory_name_en(categoryName);
         }
-        if (Category.isValidCategoryLevel(categoryLevel)) {
+        if (categoryLevel != null) {
             category.setCategory_level(categoryLevel);
         }
-        if (Category.isValidDurationBegin(durationStart)) {
-            category.setDuration_start(durationStart);
+        if (durationBegin != null) {
+            category.setDuration_begin(durationBegin);
         }
-        if (Category.isValidDurationEnd(durationEnd)) {
+        if (durationEnd != null) {
             category.setDuration_end(durationEnd);
         }
-        if (Category.isValidCategoryEnabled(categoryEnabled)) {
+        if (categoryEnabled != null) {
             category.setCategory_enabled(categoryEnabled);
         }
+
         Pagination<Category> page = new Pagination<Category>(limit, offset);
         Sort<Category> order = new Sort<Category>(orderBy, orderType, Category.class);
         page = CategoryService.getCategoryList(category, page, order);
@@ -188,7 +193,7 @@ public class CategoryResource {
             if (page.getData().size() > 0) {
                 return page;
             } else {
-                throw new NotFoundException();
+                throw new NotFoundException("Not found categories");
             }
         } else {
             throw new InternalServerErrorException();
